backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
django-cors-headers==4.0.0
python-dotenv==1.0.0
Pillow==9.5.0
django-filter==23.2
djangorestframework-simplejwt==5.2.2
celery==5.3.1
redis==4.5.5
django-celery-beat==2.5.0
reportlab==4.0.4
